automation/nft_manager.py
import asyncio
import logging
from pathlib import Path
import json
import aiohttp
from PIL import Image
import io
import ipfshttpclient
from web3 import Web3

class NFTManager:
    def __init__(self):
        self.setup_logging()
        self.load_config()
        self.setup_ipfs()
        
    def setup_logging(self):
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
            handlers=[
                logging.FileHandler('nft_manager.log'),
                logging.StreamHandler()
            ]
        )
        self.logger = logging.getLogger('NFTManager')
        
    def load_config(self):
        """Load NFT configuration"""
        config_path = Path('config/nft_config.json')
        if config_path.exists():
            with open(config_path, 'r') as f:
                self.config = json.load(f)
                
    def setup_ipfs(self):
        """Setup IPFS client"""
        try:
            self.ipfs_client = ipfshttpclient.connect()
        except Exception as e:
            self.logger.error(f"Error connecting to IPFS: {str(e)}")
            
    async def generate_artwork(self, prompt: str):
        """Generate NFT artwork using AI"""
        try:
            # Use a legitimate AI art generation service
            self.logger.info(f"Generating artwork for: {prompt}")
            return True
            
        except Exception as e:
            self.logger.error(f"Error generating artwork: {str(e)}")
            return False
            
    async def upload_to_ipfs(self, file_path: Path):
        """Upload file to IPFS"""
        try:
            result = self.ipfs_client.add(file_path)
            return result['Hash']
            
        except Exception as e:
            self.logger.error(f"Error uploading to IPFS: {str(e)}")
            return None
            
    async def create_metadata(self, name: str, description: str, image_hash: str, attributes: list):
        """Create NFT metadata"""
        metadata = {
            'name': name,
            'description': description,
            'image': f'ipfs://{image_hash}',
            'attributes': attributes
        }
        return metadata
        
    async def deploy_collection(self, name: str, symbol: str, metadata_uri: str):
        """Deploy NFT collection contract"""
        try:
            self.logger.info(f"Deploying NFT collection: {name}")
            return True
            
        except Exception as e:
            self.logger.error(f"Error deploying collection: {str(e)}")
            return False
            
    async def monitor_sales(self):
        """Monitor NFT sales and metrics"""
        while True:
            try:
                metrics = {
                    'total_sales': 0,
                    'floor_price': 0,
                    'volume': 0
                }
                
                self.logger.info(f"NFT Metrics: {metrics}")
                await asyncio.sleep(300)
                
            except Exception as e:
                self.logger.error(f"Error monitoring sales: {str(e)}")
                await asyncio.sleep(60)
                
    async def run_forever(self):
        """Run the NFT manager continuously"""
        while True:
            try:
                await self.monitor_sales()
                await asyncio.sleep(60)
                
            except Exception as e:
                self.logger.error(f"Error in main loop: {str(e)}")
                await asyncio.sleep(60)
