requirements.txt
openai==1.3.7
instructor==1.2.2
deepdiff==6.7.1
termcolor==2.3.0
python-dotenv==1.0.0
rich==13.7.0
jsonref==1.1.0
skyvern==0.2.1
openadapt==0.3.0
anthropic==0.5.0
deepseek-ai==0.1.0
firebase-admin==6.2.0
stripe==7.0.0
boto3==1.28.0
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
selenium==4.16.0
playwright==1.40.0
aiohttp==3.9.1
asyncio==3.4.3
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
python-telegram-bot==20.7
docker==7.0.0
docker-compose==1.29.2
GitPython==3.1.40
aiofiles==23.2.1
psycopg2-binary==2.9.9
telethon==1.33.1
cryptography==41.0.7
prometheus-client==0.19.0
SQLAlchemy==2.0.23
alembic==1.13.0
celery==5.3.6
redis==5.0.1
scrapy==2.11.0
undetected-chromedriver==3.5.4
python-binance==1.0.19
ccxt==4.1.13
Pillow==10.1.0
schedule==1.2.1
psutil==5.9.6
PyGithub==2.1.1
browser-cookie3==0.19.1
pycryptodome==3.19.0
web3==6.11.3
ta==0.11.0